+++ contracts/implementations/RMRKNestingWithEquippableImpl.sol
@@ -11,8 +11,11 @@ error RMRKMintUnderpriced();
 error RMRKMintZero();
 
 //Minimal public implementation of IRMRKNesting for testing.
-contract RMRKNestingWithEquippableImpl is OwnableLock, RMRKMintingUtils, RMRKNestingWithEquippable {
-
+contract RMRKNestingWithEquippableImpl is
+    OwnableLock,
+    RMRKMintingUtils,
+    RMRKNestingWithEquippable
+{
     address _equippableAddress;
 
     constructor(
@@ -22,8 +25,8 @@ contract RMRKNestingWithEquippableImpl is OwnableLock, RMRKMintingUtils, RMRKNes
         uint256 pricePerMint_,
         address equippableAddress_
     )
-    RMRKNestingWithEquippable(name_, symbol_)
-    RMRKMintingUtils(maxSupply_, pricePerMint_)
+        RMRKNestingWithEquippable(name_, symbol_)
+        RMRKMintingUtils(maxSupply_, pricePerMint_)
     {
         // Can't add an equippable deployment here due to contract size, for factory
         // pattern can use OZ clone
@@ -34,37 +37,44 @@ contract RMRKNestingWithEquippableImpl is OwnableLock, RMRKMintingUtils, RMRKNes
     Template minting logic
     */
     function mint(address to, uint256 numToMint) external payable saleIsOpen {
-        (uint nextToken, uint totalSupplyOffset) = _preMint(numToMint);
+        (uint256 nextToken, uint256 totalSupplyOffset) = _preMint(numToMint);
 
-        for(uint i = nextToken; i < totalSupplyOffset;) {
+        for (uint256 i = nextToken; i < totalSupplyOffset; ) {
             _safeMint(to, i);
-            unchecked {++i;}
+            unchecked {
+                ++i;
+            }
         }
     }
 
     /*
     Template minting logic
     */
-    function mintNesting(address to, uint256 numToMint, uint256 destinationId) external payable saleIsOpen {
-        (uint nextToken, uint totalSupplyOffset) = _preMint(numToMint);
+    function mintNesting(
+        address to,
+        uint256 numToMint,
+        uint256 destinationId
+    ) external payable saleIsOpen {
+        (uint256 nextToken, uint256 totalSupplyOffset) = _preMint(numToMint);
 
-        for(uint i = nextToken; i < totalSupplyOffset;) {
+        for (uint256 i = nextToken; i < totalSupplyOffset; ) {
             _nestMint(to, i, destinationId);
-            unchecked {++i;}
+            unchecked {
+                ++i;
+            }
         }
     }
 
-    function _preMint(uint256 numToMint) private returns(uint, uint) {
+    function _preMint(uint256 numToMint) private returns (uint256, uint256) {
         if (numToMint == uint256(0)) revert RMRKMintZero();
         if (numToMint + _totalSupply > _maxSupply) revert RMRKMintOverMax();
 
         uint256 mintPriceRequired = numToMint * _pricePerMint;
-        if (mintPriceRequired != msg.value)
-            revert RMRKMintUnderpriced();
+        if (mintPriceRequired != msg.value) revert RMRKMintUnderpriced();
 
-        uint256 nextToken = _totalSupply+1;
+        uint256 nextToken = _totalSupply + 1;
         _totalSupply += numToMint;
-        uint256 totalSupplyOffset = _totalSupply+1;
+        uint256 totalSupplyOffset = _totalSupply + 1;
 
         return (nextToken, totalSupplyOffset);
     }
@@ -78,10 +88,7 @@ contract RMRKNestingWithEquippableImpl is OwnableLock, RMRKMintingUtils, RMRKNes
         _setEquippableAddress(equippable);
     }
 
-    function transfer(
-        address to,
-        uint256 tokenId
-    ) public virtual {
+    function transfer(address to, uint256 tokenId) public virtual {
         transferFrom(_msgSender(), to, tokenId);
     }
